requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-google-genai==0.0.11
google-generativeai>=0.4.1,<0.5.0
python-dotenv==1.0.1
fastapi==0.111.1
langserve==0.2.2
sse_starlette==2.1.2
uvicorn==0.24.0
pydantic==2.5.2
python-multipart>=0.0.7 